Ayu2/backend/requirements.txt
flask==2.0.2
redis
grpcio==1.44.0
grpcio-tools==1.44.0
psycopg2-binary==2.9.3
psycopg2